Reserve register R11 for indirect jump addresses.
Rationale:

Can't we just say that call/return/jump clobbers the register it uses for
indirect jumps?

No. This is only useful before register allocation. Unfortunately
various passes insert control flow after register allocation which we
also want to rewrite. We place our rewrites at the pre-emit state. At
this point, code properties required by the register allocator no longer
exist so it's not possible to use virtual registers and have the
allocator patch it up.

OK, let's make the inserted code preserve everything.

For a jump this isn't possible since we have no way to restore the
original register value.

We *can* do this for calls, but it involves pushing the register to the
stack before the call and restoring it after the call is complete. This
should be safe, but it alters the stack layout and it would be best to
be as transparent as possible.