Define HAD_JMP with semantics that clobber EFLAGS (#3)
The previous implementation of Hadean CFI ignored the fact that
EFLAGS are clobbered by an indirect branch. We now define a custom
HAD_JMP pseudo-instruction with the correct semantics and forbid
selection of the standard JMP instructions.

We do not redefine CALL and RET despite the same applying to them,
because the calling convention states that EFLAGS are not preserved
across function calls.